rfp-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
alembic==1.12.1
httpx==0.25.2
google-generativeai==0.3.1
pypdf2==3.0.1
python-docx==1.1.0
aiofiles==23.2.1